Refactoring finished
Seems to be working, although it needs futher testing to make sure I didn't break anything :)
