+++ pogom/utils.py
@@ -11,8 +11,6 @@ import shutil
 import pprint
 import time
 import random
-import geopy
-import geopy.distance
 from s2sphere import CellId, LatLng
 
 from . import config
@@ -651,15 +649,6 @@ def in_radius(loc1, loc2, distance):
     return equi_rect_distance(loc1, loc2) < distance
 
 
-# Apply a location jitter.
-def jitter_location(location=None, maxMeters=10):
-    origin = geopy.Point(location[0], location[1])
-    b = random.randint(0, 360)
-    d = math.sqrt(random.random()) * (float(maxMeters) / 1000)
-    destination = geopy.distance.distance(kilometers=d).destination(origin, b)
-    return (destination.latitude, destination.longitude, location[2])
-
-
 def i8ln(word):
     if config['LOCALE'] == "en":
         return word
@@ -754,144 +743,3 @@ class Timer():
     def output(self):
         self.checkpoint('end')
         pprint.pprint(self.times)
-
-
-# Check if all important tutorial steps have been completed.
-# API argument needs to be a logged in API instance.
-def get_tutorial_state(api, account):
-    log.debug('Checking tutorial state for %s.', account['username'])
-    request = api.create_request()
-    request.get_player(
-        player_locale={'country': 'US',
-                       'language': 'en',
-                       'timezone': 'America/Denver'})
-
-    response = request.call().get('responses', {})
-
-    get_player = response.get('GET_PLAYER', {})
-    tutorial_state = get_player.get(
-        'player_data', {}).get('tutorial_state', [])
-    time.sleep(random.uniform(2, 4))
-    return tutorial_state
-
-
-# Complete minimal tutorial steps.
-# API argument needs to be a logged in API instance.
-# TODO: Check if game client bundles these requests, or does them separately.
-def complete_tutorial(api, account, tutorial_state):
-    if 0 not in tutorial_state:
-        time.sleep(random.uniform(1, 5))
-        request = api.create_request()
-        request.mark_tutorial_complete(tutorials_completed=0)
-        log.debug('Sending 0 tutorials_completed for %s.', account['username'])
-        request.call()
-
-    if 1 not in tutorial_state:
-        time.sleep(random.uniform(5, 12))
-        request = api.create_request()
-        request.set_avatar(player_avatar={
-            'hair': random.randint(1, 5),
-            'shirt': random.randint(1, 3),
-            'pants': random.randint(1, 2),
-            'shoes': random.randint(1, 6),
-            'avatar': random.randint(0, 1),
-            'eyes': random.randint(1, 4),
-            'backpack': random.randint(1, 5)
-        })
-        log.debug('Sending set random player character request for %s.',
-                  account['username'])
-        request.call()
-
-        time.sleep(random.uniform(0.3, 0.5))
-
-        request = api.create_request()
-        request.mark_tutorial_complete(tutorials_completed=1)
-        log.debug('Sending 1 tutorials_completed for %s.', account['username'])
-        request.call()
-
-    time.sleep(random.uniform(0.5, 0.6))
-    request = api.create_request()
-    request.get_player_profile()
-    log.debug('Fetching player profile for %s...', account['username'])
-    request.call()
-
-    starter_id = None
-    if 3 not in tutorial_state:
-        time.sleep(random.uniform(1, 1.5))
-        request = api.create_request()
-        request.get_download_urls(asset_id=[
-            '1a3c2816-65fa-4b97-90eb-0b301c064b7a/1477084786906000',
-            'aa8f7687-a022-4773-b900-3a8c170e9aea/1477084794890000',
-            'e89109b0-9a54-40fe-8431-12f7826c8194/1477084802881000'])
-        log.debug('Grabbing some game assets.')
-        request.call()
-
-        time.sleep(random.uniform(1, 1.6))
-        request = api.create_request()
-        request.call()
-
-        time.sleep(random.uniform(6, 13))
-        request = api.create_request()
-        starter = random.choice((1, 4, 7))
-        request.encounter_tutorial_complete(pokemon_id=starter)
-        log.debug('Catching the starter for %s.', account['username'])
-        request.call()
-
-        time.sleep(random.uniform(0.5, 0.6))
-        request = api.create_request()
-        request.get_player(
-            player_locale={
-                'country': 'US',
-                'language': 'en',
-                'timezone': 'America/Denver'})
-        responses = request.call().get('responses', {})
-
-        inventory = responses.get('GET_INVENTORY', {}).get(
-            'inventory_delta', {}).get('inventory_items', [])
-        for item in inventory:
-            pokemon = item.get('inventory_item_data', {}).get('pokemon_data')
-            if pokemon:
-                starter_id = pokemon.get('id')
-
-    if 4 not in tutorial_state:
-        time.sleep(random.uniform(5, 12))
-        request = api.create_request()
-        request.claim_codename(codename=account['username'])
-        log.debug('Claiming codename for %s.', account['username'])
-        request.call()
-
-        time.sleep(random.uniform(1, 1.3))
-        request = api.create_request()
-        request.mark_tutorial_complete(tutorials_completed=4)
-        log.debug('Sending 4 tutorials_completed for %s.', account['username'])
-        request.call()
-
-        time.sleep(0.1)
-        request = api.create_request()
-        request.get_player(
-            player_locale={
-                'country': 'US',
-                'language': 'en',
-                'timezone': 'America/Denver'})
-        request.call()
-
-    if 7 not in tutorial_state:
-        time.sleep(random.uniform(4, 10))
-        request = api.create_request()
-        request.mark_tutorial_complete(tutorials_completed=7)
-        log.debug('Sending 7 tutorials_completed for %s.', account['username'])
-        request.call()
-
-    if starter_id:
-        time.sleep(random.uniform(3, 5))
-        request = api.create_request()
-        request.set_buddy_pokemon(pokemon_id=starter_id)
-        log.debug('Setting buddy pokemon for %s.', account['username'])
-        request.call()
-        time.sleep(random.uniform(0.8, 1.8))
-
-    # Sleeping before we start scanning to avoid Niantic throttling.
-    log.debug('And %s is done. Wait for a second, to avoid throttle.',
-              account['username'])
-    time.sleep(random.uniform(2, 4))
-    return True
